requirements.txt
Flask==2.3.3
python-dotenv~=1.0.1
click~=8.1.7
pandas~=1.5.3